+++ src/FanController.h
@@ -27,8 +27,8 @@ struct FanControllerValues
 };
 
 void FanController_setup();
-void FanController_setHumidityAndTempFunctions(float func(), float func2());
-void FanController_setAvgHumidityAndTempFunctions(float func(), float func2());
+void FanController_setHumidityAndTempFunctions(double func(), double func2());
+void FanController_setAvgHumidityAndTempFunctions(double func(), double func2());
 void FanController_applyspeed(int volt, int id, int val);
 void FanController_processAutoControl();
 void FanController_setVoltage(int id, int min, int max);
